Make types more accurate and move creator.js to create-theme.js
