+++ src/index.d.ts
@@ -7,15 +7,16 @@ import {
 } from 'styletron-react';
 import {Override, Overrides} from './overrides';
 import {Locale} from './locale';
-import {Theme, ThemePrimitives} from './theme';
+import {Theme, ThemePrimitives, PartialThemePrimitives} from './theme';
 
 type UseStyletronFn<Theme> = () => [(arg: StyleObject) => string, Theme];
 export function createThemedUseStyletron<Theme>(): UseStyletronFn<Theme>;
 export const useStyletron: UseStyletronFn<Theme>;
 
 export function createTheme<P extends object>(
-  primitives: ThemePrimitives,
+  primitives: PartialThemePrimitives,
   overrides?: P,
+  options?: {isDark?: boolean},
 ): Theme & P;
 export function mergeOverrides<T>(
   target?: Overrides<T>,
